+++ .env.example
@@ -1,2 +1,10 @@
-DB_URI # Database conection URI
-JWT_SECRET # JWT secret password
+# Mongo connection
+DB_URI=
+# JWT secret string
+JWT_SECRET=
+# JWT secret string for refresh token
+JWT_REFRESH_SECRET=
+# URL for Swagger doc URL if RUN_MODE=PROD. If DEV, will be used default: http://localhost:3000
+API_URL=
+# Must be DEV or PROD
+RUN_MODE=DEV
