+++ app.js
@@ -6,7 +6,8 @@ import cors from "cors";
 import swaggerUi from 'swagger-ui-express';
 import swaggerDocument from './swagger.js';
 
-import usersRouter from "./routes/authRouter.js";
+import authRouter from "./routes/authRouter.js";
+import usersRouter from "./routes/userRouter.js";
 import waterRouter from "./routes/waterRouter.js";
 
 const DB_URI = process.env.DB_URI;
@@ -17,6 +18,7 @@ app.use(cors());
 app.use(express.json());
 
 app.use("/api-docs", swaggerUi.serve, swaggerUi.setup(swaggerDocument));
+app.use("/auth", authRouter);
 app.use("/users", usersRouter);
 app.use("/water", waterRouter);
