+++ schemas/userSchema.js
@@ -1,6 +1,7 @@
 import Joi from "joi";
 
 export const registerSchema = Joi.object({
+  name: Joi.string().required(),
   email: Joi.string().required(),
   password: Joi.string().required(),
   password_conform: Joi.string().required(),
